requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-decorator-include==3.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9

